Added new config option, openDataURL, that allows us to run a single instance that serves Valve3 both internally (all data exports available) and externally (only specific data exports available). Note that this parameter only has an effect if it's set. If the parameter is blank or doesn't exist at all, data exports are configured and handled in the same way as they've always been.
As a usage example, we have openDataURL set to hvointernal.wr.usgs.gov. When requests come from our internal site, that value is matched and all data exports are allowed. When a request comes from hvovalve.wr.usgs.gov, the mismatch in values means that data exports are set on a case-by-case basis as they were before.